Some Change

+++ src/time_test.cpp
@@ -1,8 +1,8 @@
 // Copyright: 2019 Mohit Saini
 // Author: Mohit Saini (mohitsaini1196@gmail.com)
 
-#include <chrono>
-#include <thread>
+#include <chrono>  // NOLINT
+#include <thread>  // NOLINT
 
 #include "gtest/gtest.h"
 #include "quick/time.hpp"
@@ -14,12 +14,12 @@ TEST(GetEpochMicroSeconds, Basic) {
 
 TEST(MicroSecondTimer, Basic) {
   qk::MicroSecondTimer timer;
-  long long t1 = timer.GetStartTime();
+  int64_t t1 = timer.GetStartTime();
   std::this_thread::sleep_for(std::chrono::milliseconds(10));
   EXPECT_GT(timer.GetElapsedTime(), 0);
   EXPECT_EQ(t1, timer.GetStartTime());
   timer.Restart();
-  long long t2 = timer.GetStartTime();
+  int64_t t2 = timer.GetStartTime();
   EXPECT_GT(t2, t1);
 }
